requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-admin-rangefilter==0.8.1
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.5.0
-e git+https://github.com/jazzband/django-silk.git@7f03cfef0b554af541dc53fc6d5bee850469098a#egg=django_silk
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
Flask==2.0.1
gprof2dot==2021.2.21
idna==3.2
inflection==0.5.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
xlrd==2.0.1
xlwt==1.3.0
